# API_documentation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: COURSE API
  description: This is the API Documentation of Course Recommender System of Team Sept-07
servers:
  - url: http://127.0.0.1:5001
...
paths:
...
  /api/signup:
    description: End Point for Signup
    post:
      tags:
        - User
      summary: For SignUp
      description: Operation to Sign up
      operationId: signup
      requestBody:
        description: SignUp
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                email:
                  type: string
                  example: abcd231@gmail.com
                password:
                  type: string
                  example: '12345678'
                name:
                  type: string
                  example: Mohd Ikram
                reset_code:
                  type: string
                  example: aaaa1111
                roll_no:
                  type: string
                  example: 21f282319
                dual_degree:
                  type: boolean
                  example: true
                side_work:
                  type: boolean
                  example: false
                additional_education:
                  type: string
                  example: B.tech
      responses:
        '201':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Account created successfully"}'
        '400':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Something Went Wrong"}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: >-
                    {"message": "The username you entered is already connected
                    to an account."}
                Example 2:
                  value: >-
                    {"message": "The email address you entered is already
                    connected to an account."}
                Example 3:
                  value: >-
                    {"message": "The roll_no you entered is already connected to
                    an account."}
  /api/forget_password:
    description: End Point for Forget Password
    post:
      tags:
        - User
      summary: For Forget Password
      description: Operation to Forget Password
      operationId: forget_password
      requestBody:
        description: Forget Password
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: mdikram88
                reset_code:
                  type: string
                  example: aaaa1111
                password:
                  type: string
                  example: '12345678'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Password Reset Successfully"}'
        '403':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "The reset code you entered is incorrect."}'
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "Account doesn''t exist"}'
...
  /api/courses/student/{user_id}:
    get:
      tags:
        - Course
      summary: For getting courses details for the student
      description: Operation to get User Course details.
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value:
                    Enrolled:
                      - - CS001214
                        - 12 weeks
                        - Someone
                        - 4
                        - 5
                        - 1
                        - Java
                    Unenrolled:
                      - - CS2001712
                        - 8 Weeks
                        - SLS
                        - 4
                        - 5
                        - 2
                        - System Commands
                    username: M.Ikram
        '404':
          description: ''
          content:
            application/json:
              examples:
                Example 1:
                  value: '{"message": "user_id is invalid"}'
...